+++ src/Freya.Machine.Extensions.Http/Constants.fs
@@ -21,7 +21,6 @@
 [<AutoOpen>]
 module internal Freya.Machine.Extensions.Http.Constants
 
-
 [<RequireQualifiedAccess>]
 module Actions =