+++ src/Freya.Machine.Extensions.Http/Decisions.fs
@@ -23,9 +23,9 @@ module internal Freya.Machine.Extensions.Http.Decisions
 
 open Freya.Core
 open Freya.Core.Operators
+open Freya.Lenses.Http
 open Freya.Machine
 open Freya.Machine.Operators
-open Freya.Types.Http
 
 (* Decisions *)