+++ src/Freya.Machine.Extensions.Http/Defaults.fs
@@ -23,8 +23,8 @@ module internal Freya.Machine.Extensions.Http.Defaults
 
 open System
 open Freya.Core
-open Freya.Types.Http
-open Freya.Types.Language
+open Arachne.Http
+open Arachne.Language
 
 (* Charsets *)