+++ tests/Freya.Machine.Extensions.Http.Tests/Domain.Tests.fs
@@ -1,8 +1,8 @@
 ﻿module Freya.Machine.Extensions.Http.Tests.Domain
 
 open Freya.Machine.Extensions.Http
-open Freya.Types.Http
-open Freya.Types.Language
+open Arachne.Http
+open Arachne.Language
 open NUnit.Framework
 open Swensen.Unquote